+++ flecsi/topo/ntree/test/ntree.cc
@@ -191,7 +191,6 @@ struct sph_ntree_t : topo::specialization<topo::ntree, sph_ntree_t> {
 using ntree_t = topo::ntree<sph_ntree_t>;
 
 sph_ntree_t::slot sph_ntree;
-sph_ntree_t::cslot coloring;
 
 const field<double>::definition<sph_ntree_t, sph_ntree_t::base::entities>
   density;
@@ -280,8 +279,8 @@ int
 ntree_driver() {
 
   std::vector<sph_ntree_t::ent_t> ents;
-  coloring.allocate("coordinates.blessed", ents);
-  sph_ntree.allocate(coloring.get(), ents);
+  sph_ntree.allocate(
+    sph_ntree_t::mpi_coloring("coordinates.blessed", ents), ents);
 
   auto d = density(sph_ntree);