+++ tutorial/4-data/3-dense.cc
@@ -8,7 +8,6 @@
 using namespace flecsi;
 
 canon::slot canonical, cp;
-canon::cslot coloring;
 
 const field<double>::definition<canon, canon::cells> pressure;
 
@@ -36,9 +35,9 @@ print(canon::accessor<ro> t, field<double>::accessor<ro> p) {
 
 void
 advance(control_policy &) {
-  coloring.allocate("test.txt");
-  canonical.allocate(coloring.get());
-  cp.allocate(coloring.get());
+  canon::mpi_coloring c("test.txt");
+  canonical.allocate(c);
+  cp.allocate(c);
 
   auto pf = pressure(canonical), pf2 = pressure(cp);