+++ tutorial/5-parallel/2-reduceall.cc
@@ -14,7 +14,6 @@
 using namespace flecsi;
 
 canon::slot canonical;
-canon::cslot coloring;
 
 const field<double>::definition<canon, canon::cells> pressure;
 
@@ -57,8 +56,7 @@ print(canon::accessor<ro> t, field<double>::accessor<ro> p) {
 
 void
 advance(control_policy &) {
-  coloring.allocate("test.txt");
-  canonical.allocate(coloring.get());
+  canonical.allocate(canon::mpi_coloring("test.txt"));
 
   auto pf = pressure(canonical);