+++ tutorial/standalone/poisson/include/state.hh
@@ -13,7 +13,6 @@ inline const field<double>::definition<mesh, mesh::vertices> sd;
 inline const field<double>::definition<mesh, mesh::vertices> Aud;
 
 inline mesh::slot m;
-inline mesh::cslot coloring;
 
 } // namespace poisson