+++ doc/sphinx/src/news.rst
@@ -33,6 +33,14 @@ Deprecated
   * ``control::execute`` |mdash| use ``control::invoke`` or ``runtime::main``
   * ``option_value`` and needing one in a ``program_option`` validation function |mdash| accept the option type instead
 
+* Data
+
+  * ``coloring_slot`` |mdash| use ``specialization::mpi_coloring``
+
+* Topologies
+
+  * ``specialization::cslot`` |mdash| use ``mpi_coloring``
+
 * Utilities
 
   * Passing binary functors to ``mpi::one_to_allv``, ``mpi::one_to_alli``, and ``mpi::all_to_allv`` |mdash| remove second parameter or use ranges
@@ -49,6 +57,7 @@ New features
 
 * Topologies
 
+  * ``specialization::mpi_coloring`` creates a coloring eagerly.
   * ``topo::make_ids<S>(r)`` is a convenience function to convert a range ``r`` of ordinary integers into a range of ``id<S>`` objects.
   * ``unstructured::special_field`` is the field definition for special-entity lists.
   * ``unstructured::get_special_entities`` allows access to individual special-entity lists.